refactor: split stats recording helpers
